Fix timeout issue with fsfreeze
We have seen several VM hangs due to a timeout during freeze. Two problems have
been observed:

1. Lack of transactionality

The snapshot code should freeze a VM while creating a snapshot if the VM is
running. The old code used to query the monitor twice for VM liveness. If the
two queries gave different answers (see below for a reason), it was possible to
have a VM frozen but never thawed.

2. Out-of-sync guest agent communication

The GA interface code tries to sync with the GA before issueing commands. If a
command take too long, the socket is closed due to a socket timeout. The next
time the GA interface opens a connection, it gets the answer to that previously
issued command after issueing a new sync request and aborts. The next time the
GA interface gets the answer to _that_ sync request while issuing a new one and
so forth.

This means that the GA interface in fc.qemu and the running monitor never catch
up after getting out of sync once. The improved code tries up to three times to
get back in sync in such a situation.

Note: unit tests are still missing.

Re #18917