Selective volume unlocking
During the burn-in test of #18771 I got half-finished migrations due to network
problems. This lead to the unfortunate case where the lock of the root volume
was already transferred, but tmp and swap were not. The cluster was not able to
recovery autonomously: the old host held tmp and swap locks forever since it did
not succeed in releasing the root lock, which was already held by the new host.

Now we release locks on a best-effort basis.

Re #18771