Refactor RBD volume handling
Extract 'Image' base class out of Volume. The add a Snapshot class to wrap
images that are actually snapshots, in contrast to Volume objects which are
always base images now.

The next step is to mount map/unmap and mount/unmout from Volume to Image so
that we can mount all kinds of images with the same API.

Re #19079