Heal inconsistent state for Ceph lock + not running here
I've seen several instances where a VM died on host A while the directory
assigns B as new VM host. Ceph locks are still active for host A. Now fc-qemu
did not do anything on either host: not on host A because the VM was not
supposed to run there anymore and not on host B because it is not able to
acquire Ceph locks.

Amend code to unlock Ceph volumes if VM is not supposed on the local host and
known not to be running, but Ceph locks are still held by the local host.

Also improve error reporting and exception handling.

Re #23695