Clean up unused resources in Agent.ensure()
If a VM is not supposed to run on the local VM host, unused resources are
cleaned up as part of the regular localconfig script. `fc-qemu ensure` takes now
care of:

- releasing lingering Ceph locks
- removing useless Consul service registrations
- cleaning up run and temp files left over by crashes.

In addition, fc-qemu also does not try to set the Qemu binary generation via GA
directly after VM startup as it was very likely to fail and the subsequent
`ensure` operation used to fail as well due to cookie sync issues.

Tests run a bit faster now since I've reduced the GA timeout for testing.

Re #23695