tests/vm_tests: increase some timeouts
The migration test appeared to be flaky recently, the logs of the failed
attempts suggest that the test just barely did not make the last
inmigrate steps before timing out.
The same is true for test_vm_resize_disk.
For now, increasing the timeouts makes the tests pass again even when
the hydra is under a bit of load.