Improve VM migration test stability.
The current test is very timing dependent. I've added a "meta" test
that we can use to validate that certain specific outputs are caught
by the Ellipsis for the migration test. If any further timing issues
in the test should occur we can add their outputs and verify that
the Ellipsis catches them all so we don't get into a "back and forth"
battle between incomplete solutions.

Fixes PL-131812