migration: fix crash due to failed heartbeats not propagating tests: reduce test flakiness
This introduces pytest-patterns as a replacement for the current
use of Ellipsis.

Increase general timeouts so tests are a bit less flaky.

Improve some lock logging to make it easier to understand and
debug when reading the locks and I think this also fixes some race
conditions.

Fix VM kill handling to include our little supervisor process and
selectively decide whether to kill it or not.

Add more testing output about the environment in case something goes
wrong in the tests and we need to debug it.

Fixes PL-131744