guest agent: revamp synchronisation and assist thawing
We think that this is the cause of PL-132809
(VMs unavailable due to hung tasks / guest freeze).

The previous way to establish a connection was broken as it
misunderstood the synchronisation mechanism and the way the guest agent
tries to send pending messages to the socket.

The new approach first tries to send a gratuitous thaw,
flush out any pending messages and then tries once to ping and
synchronize - otherwise erroring out immediately.

Also, when freezing we try harder to thaw if we failed and we increased
the freezing timeout to 5 minutes.

Fixes PL-132809