add ability perform rbd pool migration during cold boot
This uses Ceph's RBD live migration and thus is quite fast and
we continuously maintain the flattening during subsequent ensure()
calls. Existing snapshots are kept.

This includes a large refactoring to make the Ceph abstraction code
easier to understand with the different requirements we now juggle.

We basically have a relatively abstract "Ceph" object that is the
entrance point to the cluster and some convenience APIs. Volumes are
specific RBD volumes and their API and "VolumeSpecs" capture the
logic we have around Root/Tmp/Swap volumes and their details.

They get embedded into a streamlined process of
starting/ensuring/stopping so the agent and ceph manager objects
can handle them pretty "blindly". They are reachable as well known
names, though, to make it easier to handle the few special cases
in the higher layers.

re PL-131857