Manage host routes for guest interfaces in layer 3 VRFs
In order for guests interfaces which are connected to layer 3 VRFs to
be reachable, we need to manage routes installed in the VRF routing
table on the host.

This adds an extra step to the ensure process which reconciles the
state of the VRF routing table to ensure the correct routes are
assigned for guest tap interfaces. This adds a dependency on iproute2
for management of the routes, and we also provide an rt_protos file
with an fc-qemu route protocol in order make the management of routes
by fc.qemu visible in the routing table.