libceph: do not use `rbd ls` to test for image existence.
This was a mistake when changing the implementation, as this has caused
a massive load increase (that for some reason only happened after
a tipping pooint) and made a single OSD (likely due to both a
class-based instruction and a single object being responsible) the
bottleneck.

Switch to using `rbd info` instead which should avoid this hotspot.

Re PL-133632