outmigrate: improve rescue resilience to name resolution errors
We have seen twice (PL-133752, PL-133891) that name resolution errors in
our internal network cause VM migrations during that time to not only
fail, but the supposed `rescue` job then destroyed the local running VM
which would otherwise have continued to run well.

After tracking this down to an exception in the `ceph.lock` call during
rescue, I came to the conclusion that our libceph/librbd/librados
reimplementation via the ceph CLI is now much more susceptible to name
resolution errors, as each CLI command call tries to resolve mon
hostnames again and again. The real ceph lib bindings might have done
some caching or so, I did not check.

As I did not see a simple and general approach to countering that
regression, I have decided to enrich the exceptions thrown by libceph
with the information of this special type of error – a
NameResolutionError exception. Consuming code can then use this
information to implement special case handling where it matters and
is mandated – like in the Outgoing.rescue part. When we are currently
running the "original" vm and an outgoing migration fails before we even
attempted to transfer the ceph lock, we can of course still check the
volume lock status at rescue time – but if that just fails due to a mere
name resolution error, keeping the VM running is the more appropriate
reaction than destroying it.

I added some basic unit tests for crucial cases, but did not run the
whole test suite yet due to devhost oddities.
We might also want to cover this special case in a live integration
test.

Remaining tasks:
- can the ceph lock migration be shifted to *after* the actual
  migration, to widen the window where we can rescue with a clear
  conscience to "as long as the migration progress is <100%"?
- check test suite
- extend integration tests